Revert "refactor: 디렉터리 구조 변경"
This reverts commit f1b770e5a76715935f6cf663c0e8b973a95860c0.
